Clean code, move code into seperate files, refactor code.
