+++ index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Stock Analysis App</title>
-    <script type="module" src="index.js" defer></script>  
+    <script type="module" src="js/main.js" defer></script>  
   <script src="https://s3.tradingview.com/tv.js"></script>
   <script src="https://cdn.tailwindcss.com"></script>
 </head>
